Remove record field names from its data definition.
We'll have to see later, but the names probably aren't beneficial as
part of the data.  E.g. for remote logging it's going to be a lot of
duplicated information that the receiver may only need to obtain once
(or possibly receivers interpret as they wish and are left to do other
data conformity checks if they want).