Change implementation of data store expiry.
Instead of distinct timer actors tracking expiry of a given key, the
master store sends delayed reminder messages to expire a key at a given
time.  If the expiration is still the same when it comes time to expire,
the entry is deleted, else the message is ignored (because an
insert/modification occured which changed the expiration value).

Also removed RocksDB merge operator, because the current structure of
the store has separate key-spaces for values and expiries, not seeing a
good way to immediately support merge operations that need to update
both the value and the expiry time.