Adapt code base to new coding style
This involved reformatting the C and C++ code with clang-format. There
exists now a new configuration file .clang-format in the top-level
directory, which will be automatically used when invoking clang-format
any subdirectory. The following command line works well:

        clang-format -style=file -i **/*{cc?,hh?}

We've adapted the Python bindings manually. Overall, the new style
follows the principle of reducing vertical whitespace to maximize the
ability to understand the control flow in a coherent fashion.