Remove optional<T> reference specialization
The C++17 version of optional<T> does not contain a specialization for
optional<T&>, even though it was considered early on. Hence we no longer
make use of this construct in Broker.

Also, we get rid of optional<T> unit tests, since this is a CAF type.
We have added new unit tests to CAF, so the tests in Broker are now
superfluous.