Begin implementation of endpoint peering
This is the first step that implements peer management. Peerings are
uni-directional with regard to connectivity: if an endpoint loses the
connection to another, only the peering originator will attempt to
re-establish the peering. Peering are bi-directional with respect to
subscriptions: two peering endpoints exchange their subscriptions and
relay messages for each other accordingly.

This commit introduces several new types:

   - broker::network_info
        A pair of IP address and TCP port.

   - broker::node_id
        A unique ID to describe a machine and OS process combination.

   - broker::endpoint_id
        A unique ID within a node.

   - broker::endpoint_uid
        A combination of node_id and endpoint_id, enriched with network
        information when available.

   - broker::status
        A diagnostic message that endpoints receive in addition to
        general topic-message pairs.

   - broker::peer_info
        Contains information about a specific peer of an endpoint,
        including its endpoint_uid and whether there exists an
        inbound or outbound peering.

Because endpoints can either be in-memory or on a remote machine, an
endpoint has now a globally unique identifer (broker::endpoint_uid).
Note that the IP adress is not part of this information, since NATs and
other network devices may result in asymetric TCP endpoint information.