Implement data store expiration logic
This commit implements the expiration logic at the master. Users can
associate time points that represent expiration values, which are stored
along with the key-value pairs. The master gets a notification once the
time delta between the point in time of insertion and the given
expiration has elapsed. If no update to the expiration time has occurred
in the meantime, i.e., if the expiration time is greater or equal the
initially given expiration time, then the master deletes the value and
sends an erase update to all clones.