Fix Python3 unit test issue
There exists a slight difference in the way SWIG represents a void*
in Python3 vs. Python2. For version 3, __repr__ (and therefore __str__)
displays context about the SWIG wrapper object. For version 2, we get
the exact value. By "accessing" the underlying pointer value, we need to
perform an explicit cast. Since users can pass any Python object in
there, they would have to cast it back later anyway. This is what we did
in the fix: just perform the explicit cast. It worked without an
explicit cast in version 2 because the equality check (r.cookie == 42)
must have performed the cast implicitly.