Make consistent use of result<T>
We used caf::expected<T> (hoisted into namespace broker) previously as
our boxed result type. Since we got rid of broker::error in favor of
broker::status, we now simplified the API and got rid of expected<T> in
favor of broker::result<T>, which is either a T or a broker::status.