Disentangle error and status semantics
Errors and statuses were lumped together in the same structure. This
resulted in abuse of status for cases where they shouldn't be used,
e.g., in return codes for store operations. (There, the semantics are
either you get a result or an error, but certainly not a status.)

While this commit unfortunately changed again the API that we've
recently agreed upon in the mailing list, it keeps the overall code base
more maintainable.