Starting to update Broker Python bindings.
As I was looking over the bindings to understand the current state I
figured it might work best to start by disabling all the existing
pybind11 code first and then start adding it back piece by piece while
adapting it to the changed Broker C++ API as we go (and writing
tests). So I've begun that process and plan to keep working on this as
time permits. Any help is appreciagted.