Allow clones to buffer data store mutations when master is down
This adds a new argument to endpoint::attach_clone() indicating the
amount of time to buffer data store updates when the clone is
disconnected from its master.  On reconnect, any buffered mutations get
forwarded to the master.

It's still possible for updates to get lost during disconnects as the
clone fires and forgets without getting explicit acknowledgements from
the master, though this should still makes it easier to write code that
is more tolerant to transient connectivity problems.  i.e. smaller
number of updates get lost and users can rely on the builtin error
handling behavior instead of having to write one-off error handling code
everywhere they want to make changes to a data store.

In the future, it might still be possible to build in acknowledgement of
updates for better delivery guarantees (if that turns out to be a
desirable use-case).