Add 'put_unique' data store operation
This operation will insert a key-value pair into the store only if the
key did not already exist and returns a boolean data value indicating
whether the insertion took place.

This operation is unlike any previous query or update operation as it
serves both functions at the same time.  Before, an operation was either
purely an update/mutator or a query that would not change state.  Two
important notes follow from this:

1) These operations always need to be sent to the master store in order
   to query against its data and possibly update it.

2) The "pure" query operations have separate internal messaging paths
   for the async vs. sync APIs, yet 'put_unique' does not -- it goes
   through the same path as update operations (and so also benefits
   from their use of message batching, etc.).  The odd thing about
   this is that the 'put_unique' message carries along with it a request
   ID that's only relevant for its async API and just gets set to a
   dummy value for the sync API.  So, I can see in the future (e.g. if
   more hybrid operations get added) the internal messaging/API may
   benefit from distinguishing these hybrid operations from pure updates
   (e.g. internal_command -> hybrid_command).