Fix an issue that could cause message loss
Previously buffered data (e.g. due to congestion) could be incorrectly
treated as if it were sent from a remote peer and discarded upon
receiving an inbound message from that same peer.  This is due to
the filtering rule of "don't forward a message back to the sender".
The rule is valid, just not the fact that the previously buffered
message got treated as if it were received from someone else instead
of something generated locally.  Elaborated more in the code comments.