Fix python bindings unicode decode exceptions
On converting std::string data to Python string, pybind11 tries to
decode it as utf-8 and throws an exception if that fails.  Bro/Broker
transmit std::strings as containers for arbitrary bytes without any
guaranteed/known character encoding, so this commit changes the
Data.as_string() to always return a Python bytes value, but also the
Data converter now automaticaly tries to decode recieved Data strings as
utf-8 and will use Python string type (as before) if possible, else it
will leave it as the Python bytes type.