GH-85: Improve python bindings Event validity checks
This adds a valid() method to broker.zeek.Event objects that may be
called after creating them to check if the data it parsed was valid
event data.

Also extended the name() and args() methods of broker.zeek.Event to
first check for validity to avoid segfaulting.  If the validity check
fails, a ValueError exception is now thrown.