Port the flare to Windows/MSVC
The `pipe` system call only exists on POSIX. On Windows, we need to
create a pair of connected sockets. Also, file descriptors and sockets
are different things on Windows.

Fortunately, CAF already offers a set of abstractions we can use to
adjust for the platform differences. We still have to make sure to use
recv/send on Windows instead of read/write as well as calling WSAStartup
and WSACleanup.