GH-129: improve logging for stale store expiry reminders
Changed the logging of stale expiry attempts from WARNING to INFO,
since these are not unusual situations: they're expected to be possible
and the logic accounts for it when it does happen.

Also, changed the memory-backend to return "false" instead of
"no-such-key" as the later triggers an ERROR to be logged.  The reason
again is that this is not an exceptional situation.  i.e. the
abstract_backend::expire() return values have these meanings:

  * True: the key existed, was due to expire, and was removed
  * False: the key didn't exist or wasn't due to expire
  * Error: the backend encountered a problem that prevents resolving
    the call to either of the above true/false results