Make cluster-benchmark tool modular, add new mode
Rather than using flags to change the behavior, the configuration of the
tool now asks for a mode. This makes the tool easier to extend and use.
Previously the behavior of mutualy exclusive flags (for example, setting
both `--generate-config` and `--dump-stats`) was not well defined.

Furthermore, this commits adds another mode: `shrink-generator-file`.
This enables users to reduce the size of recordings retrospectively.