Disable background deserialization
Zeek configures only a few threads to the CAF thread pool. This may
introduce some heavy delays for the deserialization that takes place in
the background. Disabling this feature altogether makes sure that
messages will not end up in some queues even though it puts more
pressure on the multiplexer's thread.