Implement more lightweight origin credit tracking
The item-based tracking of message origins keeps all context for input
messages. However, benchmarking reveled that this nice-to-have property
comes at an expensive price.

The new approach for tracking how many inputs for each origin are
currently in the system leverages the reference count of Broker messages
that we need in any case rather than introducing additional state.