bindings: Rework Python for extensible metadata format
Support a keyword only metadata= argument for broker.zeek.Event() to
be either a dict or list of tuples of broker.Count() values to their
opaque data value. The count allows to identify metadata.

This also changes/enforces datetime.datetime usage for the NETWORK_TIMESTAMP
within Python which seems reasonable given that Zeek's time type is converted
into datetime.datetime rather than float in Python.

The .timestamp() accessor is gone and this isn't providing the same
convenience as there's in C++ (but one can easily subclass zeek.Event
if needed).