Implement JSON rendering for data_message
A `data_message` embeds a `data` value. According to the spec, the JSON
output contains the `@data-type` and `data` members of the embedded
`data` value at the top level, alongside the `topic` and (message)
`type` members. Hence, we need to extend the existing `encode` functions
with a (template) parameter that configures whether or not to print the
enclosing curly braces. The default behavior remains unchanged, we only
omit the curly braces from the `encode` overload for `data_message` in
order to "merge" to output of the nested `encode` call into the output.