Use new JSON encoding in the WebSocket interface
Replace the `caf::json_writer` member with a simple character buffer.
Instead of using the writer to generate JSON output, we call into the
new `format::json::v1` API. Further, add an `append_field` function to
the format API to allow us to render the simple JSON objects for ACK and
error messages directly (instead of going through a map that we then
convert to JSON).