Use the Canonical Project Structure for Broker
For the Canonical Project Structure, see
https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2018/p1204r0.html.

With the new structure, we drop the `include` and `src` sub-directories
and place the source files next to the header files under `libbroker`.
Further, we place the unit tests next to the source file with `.test.cc`
suffix.

Standalone tools have been placed alongside `libbroker` to the project
root, whereas system/integration tests as well as benchmarks remain
under `tests`, but are organized more consistently.