Add safety checks for accessing registry pointers
Add a new function pair `checked` and `checked_deref` to the `internal`
namespace that check whether a nullable type (such as `std::shared_ptr`)
contains a value before forwarding or using it. Both functions throw an
exception with a customizable error message if the check fails.

We use the new functions to make sure member variables are initialized
using a non-null value or to make sure a registry pointer is valid
before dereferencing it.