Move `id_by_value` test utility to a new fixture
The `base_fixture` is quite heavyweight and some tests, like the
`multipath` test suite, uses the fixture only for `id_by_value`. Pushing
this functionality to a new fixture reduces unnecessary dependencies on
the base fixture, allowing us to make more aggressive changes to it in
the future (or to remove it entirely).