Refactor the channel test to work without actors
The `channel` test suite relies on the low-level `base_fixture` setup
but really doesn't need actors. This new implementation simply
dispatches to producers and consumers directly from the list of outgoing
messages that we keep anyways to simulate loss rates, without needing to
go through the extra abstraction of actors.