Add new API for event observation
The new event API in Broker enables users to install custom event
observers. This commit is based on 8c85e7fb287d0af06f99f9103420cb065c4cdb4f,
part of PR #414, which introduced this notion to allow logging to be redirected
elsewhere (into Zeek). This isn't required for the backpressure backport, so I'm
removing it.