Add new trie class optimized for use as filter
The `radix_tree` made some unfortunate design decisions that mean we
cannot base a filter implementation on it. Most importantly, our topics
are stored as `string_view` in the envelopes without a null terminator.
Since the `radix_tree` requires null-terminated strings, this means we
cannot use it as-is. The class also provides several interfaces that we
do not need, such as STL-compatible iterators.

Our new `trie` implementation provides a smaller API surface and focuses
on the functionality we actually do need. Furthermore, the `trie` does
not require null-terminated strings for key lookup.