Use std::format in Broker
- re-implement the logging functions to use `std::format`
- provide formatters for Broker types implementing `convert`
- explicitly generate `to_string` instead forcing them

In particular the latter change is targeting the over-reliance on
"automagic" injection of functions which makes the code both brittle and
hard to follow. Consequently, the `std::formatter` implementations for
Broker are all opt-in via explicitly calling
`BROKER_STD_FORMATTER_IMPL`. We could try specializing based on the
existence of a `convert` function, but this could pull in unwanted types
and cause ambiguity. Instantiating the formatters via macro is more
verbose, but also safer.