Remove stale files from the repository
These files have been excluded from CMake for a couple years now. Making
the code compile again is not trivial and would require restructuring
anyways. We can always rely on Git if we want to have the old
implementation as a reference point, so no reason to have dead code
rotting in the source tree.