Add new subscription_multimap class
Currently, we have one filter per subscriber in the core actor. Then, we
ask each subscriber individually whether a certain topic matches its
filter.

The new `subscription_multimap` flips this work flow: we register all
subscribers at the topics they are interested in. Then, we make a single
pass over the subscriptions to obtain all matches via the `select`
member function. The use case we have in Broker is to have a single
filter that stores all subscribers and then do a single central lookup.
The runtime cost of `select` depends on the number of subscribed topics,
not on the number of peers. Hence, we should expect better scalability
when switching to `subscription_multimap`.