Add new dispatching mechanism, use for WebSockets
A recent benchmark revealed that Broker's CPU usage goes up
significantly with the number of peering relationships, even if the new
peers never receive any data.

The root cause for this behavior is the flow-based processing in Broker,
where each new peer results in a new observer that subscribes to the
central flow of messages. For each incoming message, its topic is then
checked against the peer's filter. Hence, each message causes N filter
lookups where N is the number of peers.

Moreover, Broker does not really utilize the feature set of CAF flows.
Ultimately, the core actor simply dispatches data by reading from
buffers filled by local producers or peers and then forwards those
messages by writing to another set of buffers that are connected to
subscribers.

The new `handler` abstraction in the core actor expresses this work flow
more aptly. A handler is simply a combination of a sink and a source. A
handler can be registered at a `subscription_multimap` for each topic it
is interested in. A handler may still ignore matched messages, for
example, a handler for a data store would ignore any message on the
topic that is not of type `command_message`.

The first component we shift over to the new handler API are the
WebSocket clients. The handlers tap into the flows, for now. However,
ultimately the flow scaffolding will be removed entirely.