Refactor push and pull callbacks on the observer
Currently, we pass the message that is entering / leaving a peer or
client buffer to the observer. However, this message isn't really useful
to an observer such as Zeek. Furthermore, this API prevents us from
batching calls to the observer. Instead of passing the message to those
callbacks, we now simply pass a `count` argument. This allows us to
"compress" multiple events and also doesn't require us to have the
messages available when calling the observer.